Bump to 10.2.0-rc1 and minor improvements
The name of the program to run from the AppImage can now be specified
by passing it as the first argument when running the AppImage. This
does not affect existing behaviour, nor does it affect APPIMAGE_EXE
usage (which overrides this anyway). See README.md for details.

The binaries inside the AppImage are now compiled with optimization
for size (-Os), as well as linker optimization (-Wl,-O1). This reduces
the size of the final AppImage by 15-20 MiB, with no noticeable impact
on performance.

The QEMU developers have bumped the minimum required GNUTLS version to
a version not provided by the OS we build on (Debian 11). As a result,
the GNUTLS dependency is now built by us as part of the AppImage build
script. The GNUTLS version we build is new enough to satisfy QEMU's
dependency, while still having the same ABI level as the one provided
by the OS. This ensures it is backwards-compatible with the other
distribution-provided libraries that link to GNUTLS. This change should
cause no noticable runtime impact, however we felt like it was worth
mentioning.